Increase key size from uint16 to uint32
This change changes the `node.keySize` field from a `uint16` to a
`uint32`. This increases the maximum key size from `64K` to `4G`.
This was possible without increasing the node size by removing the
unused `height` field.

+++ iterator.go
@@ -62,7 +62,7 @@ func (it *Iterator) Key() []byte {
 // Value returns the value at the current position.
 func (it *Iterator) Value() []byte {
 	valOffset, valSize := decodeValue(it.value)
-	return it.arena.GetBytes(valOffset, valSize)
+	return it.arena.GetBytes(valOffset, uint32(valSize))
 }
 
 // Meta returns the metadata at the current position.
